refactor: #2 - Detail 모듈 구조 리팩토링 및 뷰 파일 분리
- DetailData
  - DTO 및 RepositoryImpl을 Models, Repositorys 디렉토리로 구조화
- DetailDomain
  - Entity, Input, UseCase 파일을 Models / Repositorys / UseCases 폴더로 이동
  - 불필요한 DetailPage 파일 삭제
- DetailPresentation
  - 기존 ActionBar.swift 내부 섹션 뷰(Overview, Credit, Video, Similar, SectionHeader)를 각각 별도 파일로 분리
  - MovieDetailView 내부 구조 정리 및 여백 조정
