+++ Features/Detail/DetailData/Sources/Models/PersonDetailDTO.swift
