+++ Features/Detail/DetailData/Sources/Models/VideoDTO.swift
