--- Features/Detail/DetailDomain/Sources/DetailPage.swift
@@ -1,12 +1,0 @@
-//
-//  DetailPage.swift
-//  DetailDomain
-//
-//  Created by 백래훈 on 10/31/25.
-//
-
-import Foundation
-
-public struct DetailPage {
-    public let title: String
-}
