+++ Features/Detail/DetailDomain/Sources/Models/DetailInput.swift
