+++ Features/Detail/DetailDomain/Sources/Models/VideoEntity.swift
