+++ Features/Detail/DetailDomain/Sources/Repositorys/DetailRepository.swift
