+++ Features/Detail/DetailPresentation/Sources/Views/SimilarSection.swift
@@ -1,0 +1,53 @@
+//
+//  SimilarSection.swift
+//  DetailPresentation
+//
+//  Created by 백래훈 on 11/5/25.
+//
+
+import SwiftUI
+import DetailDomain
+import Kingfisher
+
+struct SimilarSection: View {
+    let list: [SimilarMovieEntity]
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 12) {
+            SectionHeader(title: "비슷한 작품")
+
+            ScrollView(.horizontal, showsIndicators: false) {
+                HStack(spacing: 12) {
+                    ForEach(list) { item in
+                        VStack(alignment: .leading, spacing: 6) {
+                            if let posterUrl = item.posterURL {
+                                KFImage(posterUrl)
+                                    .resizable()
+                                    .placeholder {
+                                        Color.black.opacity(0.7)
+                                        ProgressView()
+                                    }
+                                    .onFailure { error in
+                                        print("이미지 로드 실패: \(error.localizedDescription)")
+                                    }
+                                    .scaledToFill()
+                                    .frame(height: 260)
+                                    .clipped()
+                            }
+
+                            Text(item.title)
+                                .font(.caption)
+                                .lineLimit(2)
+                                .frame(width: 120, alignment: .leading)
+                        }
+                        .onTapGesture {
+                            // 네비게이션: MovieDetailView(id: item.id)
+                        }
+                    }
+                }
+                .padding(.horizontal, 16)
+            }
+        }
+        .padding(.top, 16)
+    }
+}
