--- Features/Detail/DetailPresentation/Sources/Views/bb.swift
@@ -1,8 +1,0 @@
-//
-//  Project.swift
-//  Showcase
-//
-//  Created by 백래훈 on 10/31/25.
-//
-
-import Foundation
